AddFluentValidations

+++ Service/Impl/CandidateService.cs
@@ -22,25 +22,38 @@ namespace RecruitmentSystem.Service.Impl
 
         public async Task<Candidate> GetCandidateById(int id)
         {
-            return await _candidateRepository.GetCandidateById(id);
+            var candidate = await _candidateRepository.GetCandidateById(id);
+            if (candidate == null)
+            {
+                throw new KeyNotFoundException($"Candidate with {id} not found");
+            }
+            return candidate;
         }
 
         public async Task<Candidate> AddCandidate(Candidate candidate, IFormFile resumeFile)
         {
+            if (candidate == null)
+            {
+                throw new ArgumentNullException(nameof(candidate), "Candidate details cannot be null.");
+            }
+            // **** Write validations and remove this from here
             if (string.IsNullOrWhiteSpace(candidate.RUser.Email) || string.IsNullOrWhiteSpace(candidate.RUser.Password))
             {
                 throw new ArgumentException("Email and Password are required.");
             }
 
+            // Check if email already exists
             var existingUser = await _userRepository.GetUserByEmail(candidate.RUser.Email);
             if (existingUser != null)
             {
                 throw new InvalidDataException("Email already in use");
             }
 
+            // Hash password before saving
             candidate.RUser.Password = BCrypt.Net.BCrypt.HashPassword(candidate.RUser.Password);
             candidate.RUser.Role = UserRole.Candidate;
 
+            // Save user first
             var newUser = await _userRepository.AddUser(candidate.RUser);
             candidate.RUserID = newUser.RUserID;
 
@@ -54,7 +67,7 @@ namespace RecruitmentSystem.Service.Impl
                 }
 
                 var currentDateTime = DateTime.Now.ToString("dd-MM-yyyy_HH-mm-ss");
-                var uniqueFileName = candidate.RUserID.ToString() + "_CV_"+ currentDateTime + Path.GetExtension(resumeFile.FileName);
+                var uniqueFileName = candidate.RUserID.ToString() + "_CV_" + currentDateTime + Path.GetExtension(resumeFile.FileName);
                 var filePath = Path.Combine(uploadsFolder, uniqueFileName);
 
                 using (var fileStream = new FileStream(filePath, FileMode.Create))
@@ -65,19 +78,38 @@ namespace RecruitmentSystem.Service.Impl
                 candidate.Resume = filePath;
             }
 
+            // Save Candidate
             return await _candidateRepository.AddCandidate(candidate);
         }
         public async Task<Candidate> UpdateCandidate(Candidate candidate)
         {
+            if (candidate == null || candidate.CandidateID == 0)
+            {
+                throw new ArgumentException("Invalid Candidate data.");
+            }
+
+            var existingCandidate = await _candidateRepository.GetCandidateById(candidate.CandidateID);
+            if (existingCandidate == null)
+            {
+                throw new KeyNotFoundException($"Candidate with ID {candidate.CandidateID} not found.");
+            }
+
             return await _candidateRepository.UpdateCandidate(candidate);
+
         }
 
         public async Task<bool> DeleteCandidate(int id)
         {
+
+            var isDeleted = await _candidateRepository.DeleteCandidate(id);
+            if (isDeleted == false)
+            {
+                throw new KeyNotFoundException($"Candidate with ID {id} not found.");
+            }
             return await _candidateRepository.DeleteCandidate(id);
         }
 
-         public async Task<bool> DoesCandidateExistAsync(int id)
+        public async Task<bool> DoesCandidateExistAsync(int id)
         {
             return await _candidateRepository.CandidateExistsAsync(id);
         }
